2_auto_ops/docker-workshop/backend/requirements.txt
Django==4.2
psycopg2-binary==2.9.9
django-cors-headers==4.3.1
djangorestframework==3.14.0
